actually pass the d3d11 device along so we can init

+++ tiltfive-steamvr/T5RuntimeInterface.h
@@ -17,7 +17,7 @@ private:
 	bool glassesInitialized_ = false;
 
 	double ipd_{ 0 };
-	int framebufferWidth = 1216;
-	int framebufferHeight = 768;
+	uint16_t framebufferWidth = 1216;
+	uint16_t framebufferHeight = 768;
 	float renderingFov = 48;
 };
